+++ app/models/randomization_scheme_task.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Creates a template for a task for a randomization scheme
+class RandomizationSchemeTask < ActiveRecord::Base
+  # Model Relationships
+  belongs_to :randomization_scheme
+end