+++ app/models/randomization_task.rb
@@ -1,0 +1,5 @@
+class RandomizationTask < ActiveRecord::Base
+  # Model Relationships
+  belongs_to :randomization
+  belongs_to :task, dependent: :destroy
+end