+++ test/models/randomization_task_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class RandomizationTaskTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end